requirements.txt
Django==3.1.6
Pillow==8.1.0
django-bootstrap-modal-forms==2.0.1
django-crispy==1.8.1
WhiteNoise==5.2.0
gunicorn==20.0.4